# api-spec/v1/swagger.yaml
openapi: 3.0.0
info:
  title: Work543
  description: Work543 API Server
  version: 0.5.0
servers:
  - url: http://api.work543.com/v1
    description: Main (production) server
  - url: http://localhost:3000/v1
    description: Internal staging server for testing
...
paths:
...
  /users:
    get:
      summary: 사용자 목록 반환
      description: 사용자 목록을 반환합니다. 어드민 권한이 필요합니다.
      tags:
        - Users
      responses:
        '200':    # status code
          description: A JSON array of user information
          content:
            application/json:
              schema:
                type: array
                properties:
                  items:
                    $ref: '#/components/schemas/User'
    post:
      summary: 사용자 생성
      description: 사용자를 생성합니다.
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
      responses:
        '200':    # status code
          description: A JSON of user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # end of /users
  # /users/{id}
...
  /login:
    post:
      summary: 사용자 로그인
      description: 로그인합니다.
      tags:
        - Etc
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':    # status code
          description: A JSON of user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  # end of /login
  # /logout
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firatName:
          type: string
        lastName:
          type: string
        email:
          type: string
        updatedAt:
          type: datetime
        createdAt:
          type: datetime
...